remving business logic for night opacity to a seperate file
